Expose functions
Expose functions related to the creation of collections. This allows other
macros to build them too.
